Created a constantly runnin RobotSensors Command
Created a RobotSensors Command and set it to start on robotInit, and
never end. It is the location to place the putting of data that should
always be collected onto SMD, and to place function calls that need to
be called consistently, regardless of the running commands for the
subsystem.