Setup new hopper system
Changed the hopper to the new system that has one motor stiring the
balls to prevent them packing in, and another indexing motor to keep
them entering the shooter at a regular rate.
The subsytem now has 2 functions, one for controlling each one. The
Command that used to turn all 3 of the previous motors now just turns
the stirrer motor.