Replaced LifterLegs with TurnWheelExtender
Replace the LifterLegs Subsystem with one that control a single solenoid
that will actuate a wheel down lifting up the front of the robot to make
turning easier. Removed the PNM variables from RobotMap that coresponded
to the legs that no longer exist.
Created a new command that actuates the LiftWheel, and bound it to
toggle on a button that is yet to be configured.