Setup of Chassis for CurveDrive using Heading Rate
Added if's to split the subsystem PID loop into 2 separate systems that
control if it driving straight or if it is driving in a curve, and which
way it is driving if so. The Chassis has many new constants that are the
PID defaults, with the values changeable in preferences. An enum was
added to provide control of turn direction, as well as striaght. The
setters for the left and right talons have been replaced with references
to the speed and position talons. Similarly the getError functions for
left and right have been replaced with position and speed.