Created a class to do LinearInterpolation
Created a class that does linear interpolation, taking two
ArrayList<Double>. The sanity of the inputs is checked, with a public
function that returns if the data being provided is sane.
The function GetY takes a double parameter x, and returns the y value
that corresponds, doing the interpolation on calling.