Created a subsystem for DataPoint File Handling
Created a WheelSpeedCalculations subsystem that handles uses
LinearInterp to allow for a set of data points to give speeds for any
distance. It has functions to add points, save them to file, read the
file, wipe all of them, and get a speed from a distance.

The subsystem needs to be commented, but due to time constraints that
isn't occuring immediately.