Implemented a SmartShooter
Implemented a command that takes a percentage to drive the index motors
at and an instance of CameraAim, and then runs the index motors only
when the CameraAim instance is OnTarget. When it isn't, the motors won't
run.

Adjusted CameraAim to work using 2 index motors.
Adjusted AutoShoot into AutoDumbShoot, and made it drive 2 index motors.