Gear Bar Speed Defined on Run in LowerGearPickup
Modified BasicSpinMotor to allow modification of the percentage after
construction. This change was made to allow the setting of the Gear
PercentVBus in the initialize of LowerGearPickup rather than the
constructor, so that the speed is changeable while running the code,
rather than depending on a restart to have it take effect.