Created a Continuous Drive System
A new package was made, containing a class ContinuousDrive that is the
parent of all the continuous drive commands, each a simple thing such as
turning based on distance, based on angle, or driving a distance. A
command was created to test them, by running each type in various ways.