Removed the Chassis Drive in AutoDelay
Removed all references to Chassis in AutoDelay, and ensured that doing
so wouldn't break other pieces of the code. Created a new Command,
ContDelay, which is very similar to the old AutoDelay, but modified to
extend ContinuousDrive. The locations that made use of the Chassis
functions of AutoDelay have been modified to make use of ContDelay.