Convert SplineInterpolator to use Doubles
Renamed the file to match the name of the class contained within it.
Changed the package to match its new location. Modified the file to work
with doubles instead of floats to more easily mesh with existing code.

Added a notice to the top of the file in accordance with the license it
was originally published under, notifying that the file was modified
from the original state by one other than the original publisher.