CameraAim Uses Vision Distance w/ Encoder Changes
THIS CHANGE BREAKS ALL EXISTING SHOOTER DATA FILES!

Modified CameraAim to use the groundrange from the boiler, allowing the
distance at the start to be cached, as well as the total encoder
distance. While executing, the distance being used for shooter
calibration is now the initial vision distance, plus the difference
between the current position and the cached encoder position.